Fixed bug + added the initial migration and added the skeleton of services

+++ HotelManager/HotelManager/Data/Entities/Reservation.cs
@@ -28,6 +28,6 @@ namespace HotelManager.Data.Entities
 
         public bool AllInclusive { get; set; }
 
-        public decimal Bills { get; set; }
+        public double Bills { get; set; }
     }
 }
